Add files via upload
This is a new Model: There are four neural network models in this file, open and run every code, i have tried to put some descriptions. We have one dimensional CNN without using the pretrained embedding and one Bidirectional LSTM and two others CNN(1ConvCNN). We also did a testing using the testing dataset and we go up to 70.75% for the non-pretrained model. The graphs and accuracy plots are all inside. I strongly believe this is good way to move ahead, lets discuss and agree.